requirements.txt
h5py
numpy
matplotlib
tensorflow-gpu==1.12.0
tb-nightly
pillow
future